comite_api/requirements/requirements.txt
django-filter==2.4.0
djangorestframework==3.12.2
psycopg2-binary==2.8.6
django-multiselectfield==0.1.12
uWSGI==2.0.19.1
drf-spectacular==0.13.2
djangorestframework==3.12.2
